+++ xbmc/games/controllers/types/ControllerGrid.h
@@ -1,22 +1,11 @@
 /*
- *      Copyright (C) 2017 Team Kodi
- *      http://kodi.tv
+ *  Copyright (C) 2017-2018 Team Kodi
+ *  This file is part of Kodi - https://kodi.tv
  *
- *  This Program is free software; you can redistribute it and/or modify
- *  it under the terms of the GNU General Public License as published by
- *  the Free Software Foundation; either version 2, or (at your option)
- *  any later version.
- *
- *  This Program is distributed in the hope that it will be useful,
- *  but WITHOUT ANY WARRANTY; without even the implied warranty of
- *  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
- *  GNU General Public License for more details.
- *
- *  You should have received a copy of the GNU General Public License
- *  along with this Program; see the file COPYING.  If not, see
- *  <http://www.gnu.org/licenses/>.
- *
+ *  SPDX-License-Identifier: GPL-2.0-or-later
+ *  See LICENSES/README.md for more information.
  */
+
 #pragma once
 
 #include "ControllerTree.h"
@@ -73,7 +62,7 @@ namespace GAME
     /*!
      * \brief Get the width of the controller grid
      */
-    unsigned int Width() const { return m_grid.size(); }
+    unsigned int Width() const { return static_cast<unsigned int>(m_grid.size()); }
 
     /*!
      * \brief Get the height (deepest controller) of the controller grid
