+++ xbmc/interfaces/legacy/PlayList.cpp
@@ -1,21 +1,9 @@
 /*
- *      Copyright (C) 2005-2013 Team XBMC
- *      http://kodi.tv
+ *  Copyright (C) 2005-2018 Team Kodi
+ *  This file is part of Kodi - https://kodi.tv
  *
- *  This Program is free software; you can redistribute it and/or modify
- *  it under the terms of the GNU General Public License as published by
- *  the Free Software Foundation; either version 2, or (at your option)
- *  any later version.
- *
- *  This Program is distributed in the hope that it will be useful,
- *  but WITHOUT ANY WARRANTY; without even the implied warranty of
- *  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
- *  GNU General Public License for more details.
- *
- *  You should have received a copy of the GNU General Public License
- *  along with XBMC; see the file COPYING.  If not, see
- *  <http://www.gnu.org/licenses/>.
- *
+ *  SPDX-License-Identifier: GPL-2.0-or-later
+ *  See LICENSES/README.md for more information.
  */
 
 #include "PlayList.h"
@@ -32,7 +20,7 @@ namespace XBMCAddon
   {
     //! @todo need a means to check for a valid construction
     //!  either by throwing an exception or by an "isValid" check
-    PlayList::PlayList(int playList) : 
+    PlayList::PlayList(int playList) :
       iPlayList(playList), pPlayList(NULL)
     {
       // we do not create our own playlist, just using the ones from playlistplayer
@@ -92,7 +80,7 @@ namespace XBMCAddon
           CServiceBroker::GetPlaylistPlayer().ClearPlaylist(this->iPlayList);
 
           // add each item of the playlist to the playlistplayer
-          for (int i=0; i < (int)pPlayList->size(); ++i)
+          for (int i=0; i < pPlayList->size(); ++i)
           {
             CFileItemPtr playListItem =(*pPlayList)[i];
             if (playListItem->GetLabel().empty())
