+++ xbmc/peripherals/bus/virtual/PeripheralBusAddon.h
@@ -1,32 +1,17 @@
 /*
- *      Copyright (C) 2014-2017 Team Kodi
- *      http://kodi.tv
+ *  Copyright (C) 2014-2018 Team Kodi
+ *  This file is part of Kodi - https://kodi.tv
  *
- *  This Program is free software; you can redistribute it and/or modify
- *  it under the terms of the GNU General Public License as published by
- *  the Free Software Foundation; either version 2, or (at your option)
- *  any later version.
- *
- *  This Program is distributed in the hope that it will be useful,
- *  but WITHOUT ANY WARRANTY; without even the implied warranty of
- *  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
- *  GNU General Public License for more details.
- *
- *  You should have received a copy of the GNU General Public License
- *  along with this Program; see the file COPYING.  If not, see
- *  <http://www.gnu.org/licenses/>.
- *
+ *  SPDX-License-Identifier: GPL-2.0-or-later
+ *  See LICENSES/README.md for more information.
  */
+
 #pragma once
 
 #include "addons/AddonManager.h"
-#include "addons/IAddon.h"
 #include "addons/binary-addons/BinaryAddonBase.h"
-#include "guilib/IWindowManagerCallback.h"
 #include "peripherals/PeripheralTypes.h"
 #include "peripherals/bus/PeripheralBus.h"
-#include "threads/CriticalSection.h"
-#include "utils/Observer.h"
 
 #include <memory>
 #include <string>
@@ -73,9 +58,9 @@ namespace PERIPHERALS
     PeripheralPtr GetPeripheral(const std::string &strLocation) const override;
     PeripheralPtr GetByPath(const std::string &strPath) const override;
     bool SupportsFeature(PeripheralFeature feature) const override;
-    int GetPeripheralsWithFeature(PeripheralVector &results, const PeripheralFeature feature) const override;
-    size_t GetNumberOfPeripherals(void) const override;
-    size_t GetNumberOfPeripheralsWithId(const int iVendorId, const int iProductId) const override;
+    unsigned int GetPeripheralsWithFeature(PeripheralVector &results, const PeripheralFeature feature) const override;
+    unsigned int GetNumberOfPeripherals(void) const override;
+    unsigned int GetNumberOfPeripheralsWithId(const int iVendorId, const int iProductId) const override;
     void GetDirectory(const std::string &strPath, CFileItemList &items) const override;
     void ProcessEvents(void) override;
     void EnableButtonMapping() override;
